# Add `RuleSet::Declarations` to separate concerns