--- a/lib/css_parser/rule_set.rb
+++ b/lib/css_parser/rule_set.rb
@@ -55,7 +55,6 @@ module CssParser
     def initialize(selectors, block, specificity = nil)
       @selectors = []
       @specificity = specificity
-      @declarations = {}
       @order = 0
       parse_selectors!(selectors) if selectors
       parse_declarations!(block)
@@ -66,7 +65,7 @@ module CssParser
       return '' unless property and not property.empty?
 
       property = property.downcase.strip
-      properties = @declarations.each_with_object([]) do |(key, data), val|
+      properties = declarations.each_with_object([]) do |(key, data), val|
         # puts "COMPARING #{key} #{key.inspect} against #{property} #{property.inspect}"
         val << data if key == property
       end
@@ -89,13 +88,13 @@ module CssParser
     # If the property already exists its value will be over-written.
     def add_declaration!(property, value)
       if value.nil? or value.empty?
-        @declarations.delete(property)
+        declarations.delete(property)
         return
       end
 
       property = property.downcase
       property.strip!
-      @declarations[property] = DeclarationValue.new(value, order: @order += 1)
+      declarations[property] = DeclarationValue.new(value, order: @order += 1)
     end
     alias []= add_declaration!
 
@@ -103,7 +102,7 @@ module CssParser
     #
     #  rule_set.remove_declaration!('color')
     def remove_declaration!(property)
-      @declarations.delete(property)
+      declarations.delete(property)
     end
 
     # Iterate through selectors.
@@ -126,7 +125,7 @@ module CssParser
 
     # Iterate through declarations.
     def each_declaration # :yields: property, value, is_important
-      @declarations.sort_by { |_, value| value.order }.each do |property_name, property|
+      declarations.sort_by { |_, value| value.order }.each do |property_name, property|
         yield property_name, property.value, property.important
       end
     end
@@ -167,9 +166,9 @@ module CssParser
     #
     # See http://www.w3.org/TR/CSS21/colors.html#propdef-background
     def expand_background_shorthand! # :nodoc:
-      return unless @declarations.key?('background')
+      return unless declarations.key?('background')
 
-      value = @declarations['background'].value.dup
+      value = declarations['background'].value.dup
 
       if value =~ CssParser::RE_INHERIT
         BACKGROUND_PROPERTIES.each do |prop|
@@ -184,7 +183,7 @@ module CssParser
       split_declaration('background', 'background-size', extract_background_size_from(value))
       split_declaration('background', 'background-position', value.slice(CssParser::RE_BACKGROUND_POSITION))
 
-      @declarations.delete('background')
+      declarations.delete('background')
     end
 
     def extract_background_size_from(value)
@@ -197,15 +196,15 @@ module CssParser
     # Additional splitting happens in expand_dimensions_shorthand!
     def expand_border_shorthand! # :nodoc:
       BORDER_PROPERTIES.each do |k|
-        next unless @declarations.key?(k)
+        next unless declarations.key?(k)
 
-        value = @declarations[k].value.dup
+        value = declarations[k].value.dup
 
         split_declaration(k, "#{k}-width", value.slice!(CssParser::RE_BORDER_UNITS))
         split_declaration(k, "#{k}-color", value.slice!(CssParser::RE_COLOUR))
         split_declaration(k, "#{k}-style", value.slice!(CssParser::RE_BORDER_STYLE))
 
-        @declarations.delete(k)
+        declarations.delete(k)
       end
     end
 
@@ -213,9 +212,9 @@ module CssParser
     # into their constituent parts.  Handles margin, padding, border-color, border-style and border-width.
     def expand_dimensions_shorthand! # :nodoc:
       DIMENSIONS.each do |property, (top, right, bottom, left)|
-        next unless @declarations.key?(property)
+        next unless declarations.key?(property)
 
-        value = @declarations[property].value.dup
+        value = declarations[property].value.dup
 
         # RGB and HSL values in borders are the only units that can have spaces (within params).
         # We cheat a bit here by stripping spaces after commas in RGB and HSL values so that we
@@ -245,14 +244,14 @@ module CssParser
         split_declaration(property, bottom, b)
         split_declaration(property, left, l)
 
-        @declarations.delete(property)
+        declarations.delete(property)
       end
     end
 
     # Convert shorthand font declarations (e.g. <tt>font: 300 italic 11px/14px verdana, helvetica, sans-serif;</tt>)
     # into their constituent parts.
     def expand_font_shorthand! # :nodoc:
-      return unless @declarations.key?('font')
+      return unless declarations.key?('font')
 
       font_props = {}
 
@@ -261,10 +260,10 @@ module CssParser
         font_props[prop] = 'normal'
       end
 
-      value = @declarations['font'].value.dup
+      value = declarations['font'].value.dup
       value.gsub!(%r{/\s+}, '/') # handle spaces between font size and height shorthand (e.g. 14px/ 16px)
-      is_important = @declarations['font'].important
-      order = @declarations['font'].order
+      is_important = declarations['font'].important
+      order = declarations['font'].order
 
       in_fonts = false
 
@@ -299,9 +298,9 @@ module CssParser
         end
       end
 
-      font_props.each { |font_prop, font_val| @declarations[font_prop] = DeclarationValue.new(font_val, important: is_important, order: order) }
+      font_props.each { |font_prop, font_val| declarations[font_prop] = DeclarationValue.new(font_val, important: is_important, order: order) }
 
-      @declarations.delete('font')
+      declarations.delete('font')
     end
 
     # Convert shorthand list-style declarations (e.g. <tt>list-style: lower-alpha outside;</tt>)
@@ -309,9 +308,9 @@ module CssParser
     #
     # See http://www.w3.org/TR/CSS21/generate.html#lists
     def expand_list_style_shorthand! # :nodoc:
-      return unless @declarations.key?('list-style')
+      return unless declarations.key?('list-style')
 
-      value = @declarations['list-style'].value.dup
+      value = declarations['list-style'].value.dup
 
       if value =~ CssParser::RE_INHERIT
         LIST_STYLE_PROPERTIES.each do |prop|
@@ -323,7 +322,7 @@ module CssParser
       split_declaration('list-style', 'list-style-position', value.slice!(CssParser::RE_INSIDE_OUTSIDE))
       split_declaration('list-style', 'list-style-image', value.slice!(Regexp.union(CssParser::URI_RX, /none/i)))
 
-      @declarations.delete('list-style')
+      declarations.delete('list-style')
     end
 
     # Create shorthand declarations (e.g. +margin+ or +font+) whenever possible.
@@ -341,8 +340,8 @@ module CssParser
       values = []
       properties_to_delete = []
       properties.each do |property|
-        if @declarations.key?(property) and not @declarations[property].important
-          values << @declarations[property].value
+        if declarations.key?(property) and not declarations[property].important
+          values << declarations[property].value
           properties_to_delete << property
         end
       end
@@ -350,10 +349,10 @@ module CssParser
       return if values.length <= 1
 
       properties_to_delete.each do |property|
-        @declarations.delete(property)
+        declarations.delete(property)
       end
 
-      @declarations[shorthand_property] = DeclarationValue.new(values.join(' '))
+      declarations[shorthand_property] = DeclarationValue.new(values.join(' '))
     end
 
     # Looks for long format CSS background properties (e.g. <tt>background-color</tt>) and
@@ -365,12 +364,12 @@ module CssParser
       # background-position by preceding it with a backslash. In this case we also need to
       # have a background-position property, so we set it if it's missing.
       # http://www.w3schools.com/cssref/css3_pr_background.asp
-      if @declarations.key?('background-size') and not @declarations['background-size'].important
-        unless @declarations.key?('background-position')
-          @declarations['background-position'] = DeclarationValue.new('0% 0%')
+      if declarations.key?('background-size') and not declarations['background-size'].important
+        unless declarations.key?('background-position')
+          declarations['background-position'] = DeclarationValue.new('0% 0%')
         end
 
-        @declarations['background-size'].value = "/ #{@declarations['background-size'].value}"
+        declarations['background-size'].value = "/ #{declarations['background-size'].value}"
       end
 
       create_shorthand_properties! BACKGROUND_PROPERTIES, 'background'
@@ -384,40 +383,40 @@ module CssParser
       values = []
 
       BORDER_STYLE_PROPERTIES.each do |property|
-        next unless @declarations.key?(property) and not @declarations[property].important
+        next unless declarations.key?(property) and not declarations[property].important
         # can't merge if any value contains a space (i.e. has multiple values)
         # we temporarily remove any spaces after commas for the check (inside rgba, etc...)
-        return nil if @declarations[property].value.gsub(/,\s/, ',').strip =~ /\s/
+        return nil if declarations[property].value.gsub(/,\s/, ',').strip =~ /\s/
 
-        values << @declarations[property].value
-        @declarations.delete(property)
+        values << declarations[property].value
+        declarations.delete(property)
       end
 
       return if values.empty?
 
-      @declarations['border'] = DeclarationValue.new(values.join(' '))
+      declarations['border'] = DeclarationValue.new(values.join(' '))
     end
 
     # Looks for long format CSS dimensional properties (margin, padding, border-color, border-style and border-width)
     # and converts them into shorthand CSS properties.
     def create_dimensions_shorthand! # :nodoc:
-      return if @declarations.size < NUMBER_OF_DIMENSIONS
+      return if declarations.size < NUMBER_OF_DIMENSIONS
 
       DIMENSIONS.each do |property, dimensions|
         values = %i[top right bottom left].each_with_index.with_object({}) do |(side, index), result|
-          next unless @declarations.key?(dimensions[index])
+          next unless declarations.key?(dimensions[index])
 
-          result[side] = @declarations[dimensions[index]].value
+          result[side] = declarations[dimensions[index]].value
         end
 
         # All four dimensions must be present
         next if values.size != dimensions.size
 
         new_value = values.values_at(*compute_dimensions_shorthand(values)).join(' ').strip
-        @declarations[property] = DeclarationValue.new(new_value) unless new_value.empty?
+        declarations[property] = DeclarationValue.new(new_value) unless new_value.empty?
 
         # Delete the longhand values
-        dimensions.each { |d| @declarations.delete(d) }
+        dimensions.each { |d| declarations.delete(d) }
       end
     end
 
@@ -425,26 +424,26 @@ module CssParser
     # tries to convert them into a shorthand CSS <tt>font</tt> property.  All
     # font properties must be present in order to create a shorthand declaration.
     def create_font_shorthand! # :nodoc:
-      return unless FONT_STYLE_PROPERTIES.all? { |prop| @declarations.key?(prop) }
+      return unless FONT_STYLE_PROPERTIES.all? { |prop| declarations.key?(prop) }
 
       new_value = String.new
       ['font-style', 'font-variant', 'font-weight'].each do |property|
-        unless @declarations[property].value == 'normal'
-          new_value << @declarations[property].value << ' '
+        unless declarations[property].value == 'normal'
+          new_value << declarations[property].value << ' '
         end
       end
 
-      new_value << @declarations['font-size'].value
+      new_value << declarations['font-size'].value
 
-      unless @declarations['line-height'].value == 'normal'
-        new_value << '/' << @declarations['line-height'].value
+      unless declarations['line-height'].value == 'normal'
+        new_value << '/' << declarations['line-height'].value
       end
 
-      new_value << ' ' << @declarations['font-family'].value
+      new_value << ' ' << declarations['font-family'].value
 
-      @declarations['font'] = DeclarationValue.new(new_value.gsub(/\s+/, ' '))
+      declarations['font'] = DeclarationValue.new(new_value.gsub(/\s+/, ' '))
 
-      FONT_STYLE_PROPERTIES.each { |prop| @declarations.delete(prop) }
+      FONT_STYLE_PROPERTIES.each { |prop| declarations.delete(prop) }
     end
 
     # Looks for long format CSS list-style properties (e.g. <tt>list-style-type</tt>) and
@@ -456,6 +455,8 @@ module CssParser
     end
 
   private
+
+    attr_accessor :declarations
 
     def compute_dimensions_shorthand(values)
       # All four sides are equal, returning single value
@@ -474,14 +475,14 @@ module CssParser
     def split_declaration(src, dest, value) # :nodoc:
       return unless value and not value.empty?
 
-      return if @declarations.key?(dest) && @declarations[dest].order > @declarations[src].order
+      return if declarations.key?(dest) && declarations[dest].order > declarations[src].order
 
-      @declarations[dest] = @declarations[src].dup
-      @declarations[dest].value = value
+      declarations[dest] = declarations[src].dup
+      declarations[dest].value = value
     end
 
     def parse_declarations!(block) # :nodoc:
-      @declarations = {}
+      self.declarations = {}
 
       return unless block